cleaned up repeat code, removed sinatra, ruby, SC & heroku
