--- Gemfile
@@ -1,5 +1,0 @@
-source "http://rubygems.org"
-ruby "3.1.0"
-
-gem "sinatra"
-gem "rackup"
