--- app.rb
@@ -1,10 +1,0 @@
-require "sinatra"
-
-
-get "/" do
-  erb :index
-end
-
-get "/callback.erb" do
-  erb :callback
-end
