+++ font-awesome-4.5.0/less/core.less
