+++ font-awesome-4.5.0/scss/_core.scss
