+++ style.css
@@ -12,7 +12,7 @@
     background: -ms-linear-gradient(top, #708B75, #759AAB);
     background: -o-linear-gradient(top, #708B75, #759AAB);
     background-image: -ms-linear-gradient(top, #708B75 0%, #759AAB 100%);
-/*    background-repeat: no-repeat;*/
+    background-repeat: no-repeat;
     background-size: 100%;
     -webkit-touch-callout:none;
     -webkit-user-select:none;
@@ -52,8 +52,6 @@
     box-shadow: rgba(255,255,255,0.6) 0 0 0, inset rgba(255,255,255,0.6) 0 0 3em;
   }
 
-  /*right-hand*/
-
   .theremin-right-hand {
     border: 1em solid #3D315B;
     border-radius: 1em;
@@ -63,7 +61,7 @@
     width: 125px;
     height: 570px;
     position: relative;
-    left: 10%;
+    left: 7%;
     background-color: #444B6E;
   }
 
@@ -79,9 +77,8 @@
     font-weight: bold;
     color: #FAF2A1;
     margin: 0;
-    margin-bottom: 1vw;
-    margin-left: 35.5vw;
-    margin-top: 1vw;
+    margin-left: 33vw;
+    margin-top: 0.5vw;
   }
 
   .controls-right-hand{
@@ -99,7 +96,7 @@
     text-shadow: 2px 2px 2px #3D315B;
     margin: 0;
     padding: 0;
-    margin-left: 64px;
+    margin-left: 40px;
   }
 
   .wave-form-buttons-right-hand {
@@ -107,7 +104,7 @@
     border: 0.5em solid #3D315B;
     border-radius: 1em;
     background-color: #444B6E;
-    margin-left: 3.8em;
+    margin-left: 2em;
     margin-bottom: 3em;
     padding: 5px;
   }
@@ -123,12 +120,9 @@
     border: 0.5em solid #3D315B;
     border-radius: 1em;
     background-color: #444B6E;
-    margin-left: 3.5em;
+    margin-left: 2em;
     margin-bottom: 1em;
     width: 80%;
-  /*  -webkit-box-shadow: rgba(255,255,255,0.4) 0 0 0, inset rgba(255,255,255,0.4) 0 0 1em;
-    -moz-box-shadow: rgba(255,255,255,0.4) 0 0 0, inset rgba(255,255,255,0.4) 0 0 1em;
-    box-shadow: rgba(255,255,255,0.4) 0 0 0, inset rgba(255,255,255,0.4) 0 0 1em;*/
   }
 
   #effects-buttons {
@@ -343,12 +337,12 @@
         -moz-transform: scale3d(1,1,1);
     }
   }
+
   #tonematrix-mute { 
     display: inline-block;
     font-size: 10px;
     border: 0.5em solid #3D315B;
     border-radius: 1em;
-  
   }
 
   #help {
@@ -411,58 +405,6 @@
     margin-left: 0;
   }
 
-  #soundcloud {
-    float: right;
-    margin-top: -200px;
-    margin-right: 20vw;
-    margin-bottom: 0;
-    width: 30%;
-  }
-
-  #powered-by-sc {
-    text-decoration: none;
-    font-family: 'Iceland', cursive;
-    font-size: 22px;
-    font-weight: bold; 
-    color: #FAF2A1;
-    text-shadow: 2px 2px 2px #3D315B;
-    float: left;
-  }
-
-  #powered-by-sc i {
-    color: orange;
-    padding-left: 1vw;
-  }
-  
-
-  #startRecording, #stopRecording, #playBack, #upload {
-    border: 0.5em solid #3D315B;
-    border-radius: 1em;
-    width: 12%;
-    float: left;
-  }
-
-  #soundcloud .button {
-    width: 7vw;
-  }
-
-  #sc-buttons {
-    width: 100%;
-    height: 100%;
-    float: left;
-  }
-
-  .status {
-    font-family: 'Iceland', cursive;
-    font-size: 22px;
-    font-weight: bold; 
-    color: #FAF2A1;
-    text-shadow: 2px 2px 2px #3D315B;
-    float: left;
-    padding-left: 1vw;
-    margin: 0;
-  }
-
   #modal-outer {
     position: absolute;
     top: 3vw;
@@ -510,7 +452,7 @@
   }
 
   .modal-info {
-    width: 25%;
+    width: 40%;
     margin-left: 4vw;
     float: left;
   }
@@ -534,16 +476,22 @@
     float: left;
   }
 
-  #swap {
+  #no-mobile {
     display: none;
   }
 
-@media only screen and (min-width: 1825px){
-  #soundcloud {
-    float: right;
-    margin-top: -140px;
-    margin-right: 20vw;
-    width: 30%;
+@media only screen and (max-width: 1200px){
+  #wrapper { 
+    display: none;
+  }
+
+  #no-mobile {
+    display: inline-block;
+    font-family:'Jockey One', sans-serif;
+    color: #06426c;
+    text-shadow: 2px 2px 2px grey;
+    margin-top: 4rem;
+    width: 100%;
   }
 }
