--- views/callback.erb
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head><meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-    <title>Connect with SoundCloud</title>
-  </head>
-  <body onload="window.setTimeout(window.opener.SC.connectCallback, 1)">
-    <b style="width: 100%; text-align: center;">This popup should automatically close in a few seconds</b>
-  </body>
-</html>
